requirements/base.txt
ultralytics
sympy==1.13.1
ncnn
PyYAML
